# Optimize queue API