--- a/interfaces/Plush/templates/_inc_header.tmpl
+++ b/interfaces/Plush/templates/_inc_header.tmpl
@@ -10,7 +10,7 @@
 <html>
 <head>
   <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-  <title>SABnzbd $version - $T('queued'): $mbleft $T('MB')</title>
+  <title>SABnzbd $version</title>
   <link rel="alternate" type="application/rss+xml" title="RSS 2.0" href="${path}rss?mode=history"/>
 
   <link rel="stylesheet" type="text/css" href="${path}static/stylesheets/jqueryui/overcast/jquery-ui-1.8.15.custom.css?$version"/>