--- a/sabnzbd/interface.py
+++ b/sabnzbd/interface.py
@@ -640,8 +640,10 @@ class NzoPage(object):
                 nzo_id = a
                 break
 
-        if nzo_id and NzbQueue.do.get_nzo(nzo_id):
-            info, pnfo_list, _bytespersec = build_queue_header(self.__prim, self.__web_dir)
+        nzo = NzbQueue.do.get_nzo(nzo_id)
+        if nzo_id and nzo:
+            info = build_header(self.__prim, self.__web_dir)
+            pnfo_list = [nzo.gather_info()]
 
             # /SABnzbd_nzo_xxxxx/bulk_operation
             if 'bulk_operation' in args:
@@ -828,8 +830,8 @@ class QueuePage(object):
         if not check_login():
             raise NeedLogin(self.__root, kwargs)
 
-        start = kwargs.get('start')
-        limit = kwargs.get('limit')
+        start = int_conv(kwargs.get('start'))
+        limit = int_conv(kwargs.get('limit'))
         dummy2 = kwargs.get('dummy2')
         search = kwargs.get('search')
 
@@ -1073,8 +1075,8 @@ class HistoryPage(object):
         if not check_login():
             raise NeedLogin(self.__root, kwargs)
 
-        start = kwargs.get('start')
-        limit = kwargs.get('limit')
+        start = int_conv(kwargs.get('start'))
+        limit = int_conv(kwargs.get('limit'))
         search = kwargs.get('search')
         failed_only = kwargs.get('failed_only')
         if failed_only is None: