… sabnzbd/nzbqueue.py
… sabnzbd/nzbstuff.py
… sabnzbd/osxmenu.py