While our user in #550 originally used a somewhat unfriendly tone, he pointed out a valid point: SAB's API is slow for large queues.
This triggered @shypike to take a good look at the code and realized there was great room for improvement and already [made changes that drastically improved performance](https://github.com/sabnzbd/sabnzbd/commits/feature/optimize-api).
I also did some more tweaking of those and the results are really great!

Using a queue of 600 items totalling to 1.5TB with the global queue paused, these are the timings of the ```_api_queue_default``` function averaged over 2 minutes:

```Develop``` | @shypike  fixes | + @safihre fixes
------------ | -------------  | -------------
85.58 ms | 12.31 ms | 9.24 ms
- | **6.95x** faster | **9.26x** faster

- Plush also greatly benefits, SMPL not so much because it does 3 simultaneous API requests that slow each other down.
- When pausing the queue it would normally write the whole queue to the disk, this caused a complete freeze of up to 60 seconds with such queues. Now only the article cache is saved to disk.
- Removed the option to view files of a job on the main page in SMPL, this was ugly code that shouldn't be in the back-end functions.

Surely there are more API functions that could be improved, but this is the first one to be tackled!