@Safihre 2016-04-21
CherryPy's overhead is huge.. Even though API handlers take 2 or 3 msec to return data (including JSON encoding), the whole request still takes 25msec.
Even a simple HelloWorld example of CherryPy takes 15msec to respond.

Tried Tornado (because CouchPotato uses it and it also supports IPv6 and SSL, at least in CouchPotato), it only takes 2 or 3msec for it's HelloWorld.

Maybe we should consider switching at some point?