@shypike 2016-04-21
Merged manually and pushed to my branch.