@thezoggy 2016-04-22
the sickrage guys switched from cherrypy to tornado... they regretted it. tornado is far more buggy than cherrpy and they didnt really see any speed differences that was worth anything.