@Safihre 2016-04-23
Didn't know that, only saw CP using it and the benchmarks. But indeed Google for 'sickrage tornado' gives many hits, also from this year. 

Would also be a lot of work, so maybe not worth it. Probably other things that could be optimized.